Fixed window size didn't get saved anymore
Funny. I was searching the blame in my other changes of course. Window size was saved fine on Windows, but not on Ubuntu Karmic. Debug showed that IsIconized() always returned true in the dtor of the dialog. Saving it a bit earlier in OnClose() does the trick.
Maybe something changed in GTK / Window manager / X11 ?
Strange nobody else noticed that. Probably everybody is happy with his window size and doesn't create a config from scratch (that's what set me off).

git-svn-id: http://amule.googlecode.com/svn/trunk@9893 88c6d71f-33b7-083e-b0dd-db96bf62e93a