Use precompiled headers for MSVC build
amule, amuled, amulegui projects only for now.
This reduces time for a full debug build of monolith (without libs) from 2:30 to 0:30 for me.
Changing amule.h and compile becomes fun!
Unfortunately PCH are a total loss with GCC - which could have used a little boost eve more
(being a snail in the first place). So it's MSVC only.

git-svn-id: http://amule.googlecode.com/svn/trunk@9984 88c6d71f-33b7-083e-b0dd-db96bf62e93a