Disable EC notification completely for clients not supporting it.
Just got a debug assertion from it (running amuleweb) after deleting and then readding the same file.
After deletion file doesn't get removed from m_dirty_status because packet is not sent, and then it asserts when it gets readded. 
I wonder why it was active for non-notification clients in the first place.

git-svn-id: http://amule.googlecode.com/svn/trunk@10026 88c6d71f-33b7-083e-b0dd-db96bf62e93a