Reviewed non-throwing function qualifiers in Kademlia.
We have to take special care with wxASSERT and friends, because they don't seem to
do anything like throwing an exception. They only call wxOnAssert(), which in turn
calls wxApp::OnAssert(), which we can customize - to throw an exception, like in the
unittests.

Meanwhile, fixed a vulnerability of CUInt128: there weren't proper bounds checks in
Set32BitChunk() and in SetBitNumber() in release (w/o __WXDEBUG__) builds.

git-svn-id: http://amule.googlecode.com/svn/trunk@10095 88c6d71f-33b7-083e-b0dd-db96bf62e93a