Big client change
- Introduce CClientRef class and replace all stored pointers to CUpDownClient with it.
- Each client has now a reference counter counting how many CClientRef point to it. It gets deleted only when the last ref gets unlinked. This also obsoletes the delete queue.
- The only remaining pointer is in the socket, which is unlinked reliably.
- Widely use CClientRef to wrap CUpDownClient, reduce CUpDownClient usage to core.
- amulegui has its own mini CUpDownClient class now, updownclient.h is nowhere used in it now. This also clearly shows what is still missing (to be fixed soon).

There ARE zombie clients now, but at least they are live and valid zombies that don't crash the app. Putting them to rest remains to be done. We are hopefully better now than before, and I dearly need a snapshot.

I'm not happy about such a sweeping change when things should be getting stable. But we are having deleted-client-crashes, and this should rip out the root of the problem.

git-svn-id: http://amule.googlecode.com/svn/trunk@10401 88c6d71f-33b7-083e-b0dd-db96bf62e93a