Get rid of the strict-aliasing warnings with GCC 4 and optimizations enabled
The check is only for the GNU C++ compiler, because that will use builtin memcpy()
with -O2 (which also makes strict aliasing enabled), and thus it won't cause any
runtime penalty.  To actually get a call to memcpy() you need to disable optimization
and manually enable strict aliasing ;)

Also the check for the GNU C++ library has been disabled because its result is not used.

git-svn-id: http://amule.googlecode.com/svn/trunk@10703 88c6d71f-33b7-083e-b0dd-db96bf62e93a