Allow any kind of file links in eMuleCollection files
Removed duplicate link checking, only basic checks are applied on the link.
aMule will check each link thoroughly before accepting it.

For text collections, we pass each link to aMule as-is, keeping all fields
that might be present. For binary collections, we build a text link, adding
the AICH root hash if present in the binary.


git-svn-id: http://amule.googlecode.com/svn/trunk@10834 6ea3b165-b8b8-dda6-0cea-db2f03e7d5ff